raw-search/requirements.txt
openai
pandas
numpy
pinecone-client
tiktoken
tqdm
anthropic
langchain